Simplify the results.go/stats.go files
